+++ src/main/java/io/mcmaster/monkeypatches/MonkeyPatches.java
@@ -4,10 +4,17 @@ import org.slf4j.Logger;
 
 import com.mojang.logging.LogUtils;
 
+import io.mcmaster.monkeypatches.fixes.createstuffadditions.fluid_capability.FluidCapability;
+import net.mcreator.createstuffadditions.init.CreateSaModItems;
 import net.neoforged.bus.api.IEventBus;
+import net.neoforged.bus.api.SubscribeEvent;
 import net.neoforged.fml.common.Mod;
 import net.neoforged.fml.config.ModConfig;
+import net.neoforged.neoforge.capabilities.Capabilities;
+import net.neoforged.neoforge.capabilities.RegisterCapabilitiesEvent;
+import net.neoforged.neoforge.common.NeoForge;
 import net.neoforged.fml.ModContainer;
+import net.neoforged.fml.ModList;
 import net.neoforged.fml.event.lifecycle.FMLCommonSetupEvent;
 
 @Mod(MonkeyPatches.MODID)
@@ -16,14 +23,50 @@ public class MonkeyPatches {
     public static final Logger LOGGER = LogUtils.getLogger();
 
     public MonkeyPatches(IEventBus modEventBus, ModContainer modContainer) {
-        modEventBus.addListener(this::commonSetup);
+        // modEventBus.addListener(this::commonSetup);
+        // modEventBus.addListener(this::registerCapabilities);
+        modEventBus.register(this);
 
         // NeoForge.EVENT_BUS.register(this);
 
         modContainer.registerConfig(ModConfig.Type.COMMON, Config.SPEC);
     }
 
+    @SubscribeEvent
     private void commonSetup(FMLCommonSetupEvent event) {
         io.mcmaster.monkeypatches.util.PatchInfo.logPatchStatus();
+    }
+
+    @SubscribeEvent
+    private void registerCapabilities(RegisterCapabilitiesEvent event) {
+        /**
+         * Register Create Stuff & Additions fluid handler capabilities
+         * 
+         * Checks whether Create Stuff 'N Additions CF59 capabilities should be
+         * applied.
+         */
+
+        if (ModList.get().isLoaded("create_sa") && Config.PatchConfig.isCreateStuffAdditionsCF59Enabled()) {
+            // CreateSaModItems.ANDESITE_JETPACK_CHESTPLATE
+            FluidCapability.registerCapability(event, CreateSaModItems.ANDESITE_JETPACK_CHESTPLATE.get());
+            // CreateSaModItems.COPPER_JETPACK_CHESTPLATE
+            FluidCapability.registerCapability(event, CreateSaModItems.COPPER_JETPACK_CHESTPLATE.get());
+            // CreateSaModItems.BRASS_JETPACK_CHESTPLATE
+            FluidCapability.registerCapability(event, CreateSaModItems.BRASS_JETPACK_CHESTPLATE.get());
+            // CreateSaModItems.NETHERITE_JETPACK_CHESTPLATE
+            FluidCapability.registerCapability(event, CreateSaModItems.NETHERITE_JETPACK_CHESTPLATE.get());
+            // CreateSaModItems.SMALL_FILLING_TANK
+            FluidCapability.registerCapability(event, CreateSaModItems.SMALL_FILLING_TANK.get());
+            // CreateSaModItems.SMALL_FUELING_TANK
+            FluidCapability.registerCapability(event, CreateSaModItems.SMALL_FUELING_TANK.get());
+            // CreateSaModItems.MEDIUM_FILLING_TANK
+            FluidCapability.registerCapability(event, CreateSaModItems.MEDIUM_FILLING_TANK.get());
+            // CreateSaModItems.MEDIUM_FUELING_TANK
+            FluidCapability.registerCapability(event, CreateSaModItems.MEDIUM_FUELING_TANK.get());
+            // CreateSaModItems.LARGE_FILLING_TANK
+            FluidCapability.registerCapability(event, CreateSaModItems.LARGE_FILLING_TANK.get());
+            // CreateSaModItems.LARGE_FUELING_TANK
+            FluidCapability.registerCapability(event, CreateSaModItems.LARGE_FUELING_TANK.get());
+        }
     }
 }